HID-1965: comment an unexplainable quirk of email confirmation
I struggled for hours trying to figure out why the hash params
we were sending out were invalid once the email was clicked. It
boiled down to the missing emailId param which was not being
sent to the hash verification function.

Unfortunately from my POV, this should have never worked, but
it does work on prod. So I left a comment to maybe one day
re-investigate and determine what is going on.