HID-2184: less definitive messaging around password resets
We want to make it clear the form submitted ok, but our internal
business logic is dependent on a couple conditions, one of which
is a security hole to disclose as user feedback within the message.

Therefore, we need to make it clear that the message _probably_
gets sent to your address, instead of implying it already got sent.