[OPS-7830] Turns out the docker env is bassed as a tring boolean with a capital T, so check for that.
Turns out `{}` works fine as a TLS config object with the AWS certificate and chain. It may not work so well with self-signed certs.

Also, it does not work at all with the conditional I added. When set to `null`, the tls config will cause the API to just outright throw 502 Nope. So it will need work, possibly a conditional cache config export based on the redis TLS env variable. Yay, wheee!