chore: update some dependencies that npm 9 disliked
Peer deps behavior is stricter and they all had to be
upgraded in one go unless we used flags and I didn't
feel like researching to see the consequences.

Linting seems to work, but E2E didn't when I committed
this change. More poking or perhaps a follow-up ticket
is required to get E2E back on track.

Refs: HID-2382